test for historical and incremental poll
