+++ DataConnectors/digital shadows/DigitalShadowsSearchlight_API_functionApp.json
@@ -2,7 +2,7 @@
     "id": "DigitalShadowsSearchlightAzureFunctions",
     "title": "Digital Shadows Searchlight",
     "publisher": "Digital Shadows",
-    "descriptionMarkdown": "The Digital Shadows data connector provides ingestion of the incidents and alerts from Digital Shadows Searchlight into the Azure Sentinel using the REST API. The connector will provide the incidents and alerts information such that it helps to examine, diagnose and analyse the potential security risks and threats with efficiency.",
+    "descriptionMarkdown": "The Digital Shadows data connector provides ingestion of the incidents and alerts from Digital Shadows Searchlight into the Azure Sentinel using the REST API. The connector will provide the incidents and alerts information such that it helps to examine, diagnose and analyse the potential security risks and threats.",
     "graphQueries": [
         {
             "metricName": "Total data received",
@@ -83,7 +83,7 @@
         },
         {
             "title": "",
-            "description": "**STEP 2 - Choose ONE from the following two deployment options to deploy the connector and the associated Azure Function**\n\n>**IMPORTANT:** Before deploying the 'Digital Shadows Searchlight' connector, have the Workspace ID  and Workspace Primary Key (can be copied from the following), as well as the 'Digital Shadows Searchlight' API authorization key, secret and account ID, readily available.",
+            "description": "**STEP 2 - Choose ONE from the following two deployment options to deploy the connector and the associated Azure Function**\n\n>**IMPORTANT:** Before deploying the 'Digital Shadows Searchlight' connector, have the Workspace ID  and Workspace Primary Key (can be copied from the following), as well as the 'Digital Shadows Searchlight' API authorization key(s) or Token, readily available.",
             "instructions": [
                 {
                     "parameters": {
