--- DataConnectors/digital shadows/TimerTrigger1/AS_api.py
@@ -1,54 +1,0 @@
-""" handles all Azure sentinel apis related functions here """
-import base64
-import datetime
-import requests
-import hmac
-import hashlib
-
-""" class for log analytics api """
-class logs_api:
-
-    """ constructor initializing azure creds """
-    """ id is workspace id and key is primary key """
-    def __init__(self, id, key):                            
-        self.customer_id = id
-        self.shared_key = key
-
-    """ Build the API signature """
-    def build_signature(customer_id, shared_key, date, content_length, method, content_type, resource):
-        x_headers = 'x-ms-date:' + date
-        string_to_hash = method + "\n" + str(content_length) + "\n" + content_type + "\n" + x_headers + "\n" + resource
-        bytes_to_hash = bytes(string_to_hash, encoding="utf-8")  
-        decoded_key = base64.b64decode(shared_key)
-        encoded_hash = base64.b64encode(hmac.new(decoded_key, bytes_to_hash, digestmod=hashlib.sha256).digest()).decode()
-        authorization = "SharedKey {}:{}".format(customer_id,encoded_hash)
-        return authorization
-
-    """ Build and send a request to the POST API """
-    def post_data(self, body, log_type):
-        method = 'POST'
-        content_type = 'application/json'
-        resource = '/api/logs'
-        rfc1123date = datetime.datetime.utcnow().strftime('%a, %d %b %Y %H:%M:%S GMT')
-        content_length = len(body)
-        signature = logs_api.build_signature(self.customer_id, self.shared_key, rfc1123date, content_length, method, content_type, resource)
-        uri = 'https://' + self.customer_id + '.ods.opinsights.azure.com' + resource + '?api-version=2016-04-01'
-
-        headers = {
-            'content-type': content_type,
-            'Authorization': signature,
-            'Log-Type': log_type,
-            'x-ms-date': rfc1123date
-        }
-
-        response = requests.post(uri,data=body, headers=headers)
-        print(response.text)
-        if (response.status_code >= 200 and response.status_code <= 299):
-            print('Accepted')
-        else:
-            print("Response code: {}".format(response.status_code))
-
-
-""" class for api modifiying and getting the data in incidents section of azure sentinel using management api """
-class management_api:
-    pass
