--- DataConnectors/digital shadows/TimerTrigger1/AS_poller.py
@@ -1,4 +1,0 @@
-""" polls data from azure sentinel incidents to DS """
-
-class poller:
-    pass
