+++ DataConnectors/digital shadows/requirements.txt
@@ -3,5 +3,5 @@
 # Manually managing azure-functions-worker may cause unexpected issues
 
 azure-functions
-azure-storage-file-share==12.3.0
+azure-storage-file-share
 requests
